requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic[email]>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.0.3
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2  # Required for FastAPI TestClient
python-multipart>=0.0.6  # Required for form data handling
requests
jinja2
